VITIS-8049 Refactor SubCmdExamine to use common file
