+++ src/runtime_src/core/tools/xbutil2/SubCmdExamine.h
@@ -18,11 +18,9 @@
 #define __SubCmdExamine_h_
 
 #include "tools/common/SubCmd.h"
+#include "tools/common/SubCmdExamineInternal.h"
 
-class SubCmdExamine : public SubCmd {
- public:
-  virtual void execute(const SubCmdOptions &_options) const;
-
+class SubCmdExamine : public SubCmdExamineInternal {
  public:
   SubCmdExamine(bool _isHidden, bool _isDepricated, bool _isPreliminary);
